Use bookmarks created after the latest snapshot
This commit changes syncoid's behavior so it is always looking for
a matching snapshot and a matching bookmark. If the bookmark was created
after the snapshot it is used instead. This allows replication when the
latest snapshot replicated was deleted on the source, a common snapshot
was found but rollback on the target is not allowed. The matching bookmark
is used instead for replication.

This fixes https://github.com/jimsalterjrs/sanoid/issues/602

Signed-off-by: Felix Matouschek <felix@matouschek.org>